+++ changelog.md
@@ -1,2 +1,13 @@
+# 1.0.0-beta.2
+
+**New Features**
+- Option to delete levels after you're done with them
+- Added a button on the Lobby screen that links to the Discord server
+
+**Bugfixes**
+- Fixed an issue on Wine where certificates weren't properly created
+- Fixed weird swap behavior after someone disconnects
+- Fixed the issue where the game crashes if you are not logged in
+
 # 1.0.0-beta.1
 - Initial release of Creation Rotation